Updated PaisService and other minor changes

+++ src/main/java/main/controllers/ControllerGenerico.java
@@ -47,7 +47,7 @@ public class ControllerGenerico<E, S extends IservicioGenerico<E>> {
 
     @GetMapping("/{id}")
     @Transactional
-    public ResponseEntity<?> getOne(@PathVariable int id) {
+    public ResponseEntity<?> getOne(@PathVariable long id) {
 
         try {
 
@@ -81,7 +81,7 @@ public class ControllerGenerico<E, S extends IservicioGenerico<E>> {
 
     @PutMapping("/{id}")
     @Transactional
-    public ResponseEntity<?> put(@PathVariable int id, @RequestBody E personaForm) {
+    public ResponseEntity<?> put(@PathVariable long id, @RequestBody E personaForm) {
 
         try {
 
@@ -97,7 +97,7 @@ public class ControllerGenerico<E, S extends IservicioGenerico<E>> {
 
     @DeleteMapping("/{id}")
     @Transactional
-    public ResponseEntity<?> delete(@PathVariable int id) {
+    public ResponseEntity<?> delete(@PathVariable long id) {
 
         try {
